backend/requirements.txt
pandas
tweepy
alembic==0.8.9
bcrypt==3.1.2
click==6.6
coverage==4.2
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
psycopg2-binary
pycparser==2.17
PyJWT==1.4.2
python-editor==1.0.3
six==1.10.0
SQLAlchemy==1.1.4
Werkzeug==0.11.13
